# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TK 多站点达人工作台 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "邮箱密码登录",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/markets": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "站点目录",
        "operationId": "markets_index_api_v1_markets_get",
        "parameters": [
          {
            "name": "region_group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "europe",
              "title": "Region Group"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "tkws_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tkws Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Market"
                  },
                  "title": "Response Markets Index Api V1 Markets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant_id}/system/jobs": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "创建测试任务（幂等）",
        "operationId": "create_job_api_v1_tenants__tenant_id__system_jobs_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "tkws_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tkws Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "system"
        ],
        "summary": "我的任务",
        "operationId": "list_jobs_api_v1_tenants__tenant_id__system_jobs_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "tkws_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tkws Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobView"
                  },
                  "title": "Response List Jobs Api V1 Tenants  Tenant Id  System Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant_id}/system/checks": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "外部服务连通检查（仅管理员）",
        "operationId": "external_checks_api_v1_tenants__tenant_id__system_checks_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "tkws_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tkws Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalChecks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateJobRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "system.noop",
            "title": "Kind",
            "default": "system.noop"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "title": "CreateJobRequest"
      },
      "ExternalCheck": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "not_configured",
              "error"
            ],
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "available_credits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Credits"
          },
          "credit_cost": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Cost"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ExternalCheck"
      },
      "ExternalChecks": {
        "properties": {
          "llm": {
            "$ref": "#/components/schemas/ExternalCheck"
          },
          "fastmoss": {
            "$ref": "#/components/schemas/ExternalCheck"
          }
        },
        "type": "object",
        "required": [
          "llm",
          "fastmoss"
        ],
        "title": "ExternalChecks"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "waiting_input",
              "cancel_requested",
              "succeeded",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "last_error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "stage",
          "progress",
          "result",
          "last_error",
          "created_at",
          "updated_at",
          "finished_at"
        ],
        "title": "JobView"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+$",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "Market": {
        "properties": {
          "market_code": {
            "type": "string",
            "title": "Market Code"
          },
          "name_zh": {
            "type": "string",
            "title": "Name Zh"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "region_group": {
            "type": "string",
            "title": "Region Group"
          },
          "settlement_currency": {
            "type": "string",
            "title": "Settlement Currency"
          },
          "default_time_zone": {
            "type": "string",
            "title": "Default Time Zone"
          },
          "content_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Languages"
          },
          "fastmoss_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fastmoss Region"
          },
          "data_status": {
            "type": "string",
            "enum": [
              "queryable",
              "queryable_currency_unknown",
              "platform_unconfirmed",
              "manual_import_only"
            ],
            "title": "Data Status"
          },
          "last_probe_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Probe Total"
          },
          "probe_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probe Currency"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "market_code",
          "name_zh",
          "name_en",
          "region_group",
          "settlement_currency",
          "default_time_zone",
          "content_languages",
          "fastmoss_region",
          "data_status",
          "last_probe_total",
          "probe_currency",
          "notes"
        ],
        "title": "Market"
      },
      "Me": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantMembership"
              },
              {
                "type": "null"
              }
            ]
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/TenantMembership"
            },
            "type": "array",
            "title": "Memberships"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "display_name",
          "email",
          "current",
          "memberships"
        ],
        "title": "Me"
      },
...
      "TenantMembership": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "tenant_name": {
            "type": "string",
            "title": "Tenant Name"
          },
          "principal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "bd",
              "viewer"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "tenant_name",
          "principal_id",
          "role"
        ],
        "title": "TenantMembership"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}